requirements.txt
pandas
numpy
faker
scikit-learn
xgboost
seaborn
matplotlib
mysql-connector-python
